% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-pkg.R
\name{.build_internal}
\alias{.build_internal}
\title{build the internal package}
\usage{
.build_internal(.pkgsdir = file.path(here::here(), "packages"))
}
\arguments{
\item{.pkgsdir}{path to packages dir where internal/ and bin/ live}
}
\description{
build the internal package
}
