% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bookdown.R
\name{infuse_files}
\alias{infuse_files}
\title{infuse files or get the parameters needed}
\usage{
infuse_files(files, ..., .return_opts = F)
}
\arguments{
\item{files}{vector of files to infuse}

\item{...}{parameters to pass to infuse}

\item{.return_opts}{return the requested variables instead of infusing}
}
\value{
list
}
\description{
infuse files or get the parameters needed
}
