% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_overseer.R
\name{init_overseer}
\alias{init_overseer}
\title{initialize an overseer modeling folder}
\usage{
init_overseer(dir = "modeling", abs = FALSE)
}
\arguments{
\item{dir}{directory name}

\item{abs}{whether the directory is}
}
\description{
initialize an overseer modeling folder
}
