% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed-packages.R
\name{installed_packages_metadata}
\alias{installed_packages_metadata}
\title{get information about installed packages}
\usage{
installed_packages_metadata()
}
\description{
get information about installed packages
}
\examples{
\dontrun{
library(purrr)
# delete all remotely installed packages
ipm <- installed_packages_metadata()
ipm \%>\%
 walk(~ unlink(.x$path, recursive = TRUE))
}
}
