% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkdirp}
\alias{mkdirp}
\title{recursively create a dir}
\usage{
mkdirp(.dir, verbose = FALSE)
}
\arguments{
\item{.dir}{path}

\item{verbose}{give a message about dir creation}
}
\description{
recursively create a dir
}
