% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os_line_ending.R
\name{os_line_break}
\alias{os_line_break}
\title{os specific line ending}
\usage{
os_line_break(force_value = NULL)
}
\arguments{
\item{force_value}{force a specific return value, to override the platform type}
}
\description{
os specific line ending
}
\details{
returns a carriage return \code{\\r\\n} for windows and \code{\\n} for unix
}
