% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permit_mutation}
\alias{permit_mutation}
\title{Choose how extensively devutils affects a package namespace}
\usage{
permit_mutation(which = c("all", "some", "none"), env = parent.frame())
}
\description{
\code{devutils} will invariably inject some objects into the package namespace.
However, the extent to which this happens may be undesirable. Unless new
privileges are granted, \code{devutils} will aim to minimally affect your
namespace.
}
\section{Permission Tiers}{

Each level of permissioning grants additional permissions to modify your
package namespace. From least to most extensive, permissions are granted as
one of

\describe{
\item{"none"}{
Minimalist permissions, and the default behavior.
\itemize{
\item{Creates a \code{.devutils} environment in your package namespace
wherein settings and data are stored}
}
}
\item{"some"}{
Moderate permissions, modifying your namespace, but only as a side effect
of direct use of \code{devutils} functions.
\itemize{
\item{using \code{suggested} creates an object of the package's name in the
package environment, which can be used with \code{$} or \code{[[} indexing}
}
}
\item{"all"}{
Full permissions, additionally masking some base functions in your
package namespace.
\itemize{
\item{Masks base \code{::} and \code{:::} to automatically handle suggested
packages, throwing errors when unavailable suggested package functions
are called}
}
}
}
}

\examples{
\dontrun{permit_mutation("some")
utils <- suggested("utils")
utils$head(1:10)

permit_mutation("all")
suggested("utils")
utils::head(1:10)

permit_mutation("all")
suggested("not.a.real.package")
not.a.real.package::head(1:10)
# Error: This feature is unavailable because package 'not.a.real.package'
#   is not installed.
}

}
