% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pkgoption_fns}
\alias{pkgoption_fns}
\alias{pkgoption_fn_parse}
\alias{pkgoption_fn_is_true}
\alias{pkgoption_fn_is_false}
\alias{pkgoption_fn_str_split}
\title{Utility functions for processing option values from strings}
\usage{
pkgoption_fn_parse(...)

pkgoption_fn_is_true(...)

pkgoption_fn_is_false(...)

pkgoption_fn_str_split(delim = ";", ...)
}
\description{
Utility functions for processing option values from strings
}
