% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R, R/reload_internal.R
\name{reload_internal}
\alias{reload_internal}
\title{reload internal package}
\usage{
reload_internal()

reload_internal()
}
\arguments{
\item{...}{arguments to pass to \code{\link[=reload_namespace]{reload_namespace()}}}
}
\description{
reload internal package

reload the internal package
}
