% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_default_description.R
\name{update_default_description}
\alias{update_default_description}
\title{update the default decription file a new package}
\usage{
update_default_description(title, description, name = whoami::fullname(),
  email = whoami::email_address(), .dp = "DESCRIPTION", write = TRUE)
}
\arguments{
\item{title}{title of package}

\item{description}{description of package}

\item{name}{name}

\item{email}{author email}

\item{.dp}{description path}

\item{write}{write out the new description}
}
\value{
desc file
}
\description{
update the default decription file a new package
}
