% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bookdown.R
\name{use_bookdown}
\alias{use_bookdown}
\title{use bookdown in a folder}
\usage{
use_bookdown(author, description, ..., .path = ".", .see_opts = FALSE)
}
\arguments{
\item{author}{author of bookdown}

\item{description}{description of project}

\item{...}{optional values to pass to infuser}

\item{.path}{path to place bookdown template}

\item{.see_opts}{just print out the optional values to infuse}
}
\description{
use bookdown in a folder
}
