% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_internal.R
\name{use_internal}
\alias{use_internal}
\title{use internal packages setup in a projet}
\usage{
use_internal(proj, name = whoami::fullname(),
  email = whoami::email_address(), pkg_dir = "./packages")
}
\arguments{
\item{proj}{project name}

\item{name}{name}

\item{email}{email address}

\item{pkg_dir}{directory for internal project packages}
}
\description{
use internal packages setup in a projet
}
\examples{
\dontrun{
# autodetection of user information
use_internal("Test Project")

# manual specification
use_internal("Test Project", "Devin Pastoor", "devin.pastoor@gmail.com")
}
}
