% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal_description.R
\name{create_internal_desc}
\alias{create_internal_desc}
\title{create decription file for internal package}
\usage{
create_internal_desc(proj, first_name, last_name, email)
}
\arguments{
\item{proj}{project name}

\item{first_name}{author first name}

\item{last_name}{author last name}

\item{email}{author email}
}
\value{
desc file
}
\description{
create decription file for internal package
}
\details{
creates the desc object representation with the fields
needed to bootstrap an internal package description
}
