% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggested.R
\name{suggested}
\alias{suggested}
\title{Import a package as a Suggests dependency}
\usage{
suggested(
  pkgname,
  pkgversion = "",
  unavailable_msg = NULL,
  env = parent.frame()
)
}
\arguments{
\item{pkgname}{A \code{character} package name to import as a "Suggests"
dependency}

\item{pkgversion}{A \code{character} vector of package version
specifications, as they would be specified in \code{DESCRIPTION} (for
example, \code{">= 1.2.3"}).}

\item{unavailable_msg}{A \code{character} string to display if the suggested
package dependency is not available.}

\item{env}{A \code{environment} in which the suggested package object should
be assigned. By default, an object of the same name as the package is
created in the parent environment.}
}
\description{
Creates an object in the current environment with the same name as the
package, which can be used just like a package to access its namespace,
throwing an error if the package is not installed.
}
