% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/use_internal.R
\name{use_internal}
\alias{use_internal}
\title{use internal packages setup in a projet}
\usage{
use_internal(proj, first_name, last_name, email, pkg_dir = "./packages")
}
\arguments{
\item{.dir}{directory for internal project packages}
}
\description{
use internal packages setup in a projet
}
